seperated out the header code from the Dashboard and Dashboard-styling files to make it have its own file/component and styling file.
