+++ src/components/Dashboard/styles/UserMenu-styling.js
@@ -7,11 +7,16 @@ export const StyledDiv = styled.div`
   .userInfo {
     display: flex;
     flex-direction: column;
-    padding: 0 30px;
+    padding: 0 15px;
     padding-top: 15px;
     font-size: 1.8rem;
     border-bottom: 3px solid var(--background);
     margin-bottom: 11px;
+
+    .icon {
+      width: 65px;
+      margin: 0 auto;
+    }
 
     h3 {
       margin: 15px 0;
